Cors eingebaut

+++ api-server/src/Routes/exercise.route.ts
@@ -3,6 +3,7 @@ import express from 'express';
 import { Exercise } from '../models/exercise';
 import { GenericDAO } from '../models/generic.dao';
 import { MongoGenericDAO } from '../models/mongo-generic.dao';
+import { authService } from './services/auth.service.js';
 
 const router = express.Router();
 
@@ -27,20 +28,44 @@ router.get('/exercises', async (req, res) => {
   }
 });
 
-router.post('/exercises', (req, res) => {
+router.post('/exercises', authService.authenticationMiddleware, async (req, res) => {
   try {
     const exerciseDAO: GenericDAO<Exercise> = req.app.locals.exerciseDAO;
-    const exercise = exerciseDAO.create(req.body);
-    res.status(201).send(exercise);
+    const exercise = await exerciseDAO.create({
+      userId: res.locals.user.id,
+      name: req.body.name,
+      description: req.body.description,
+      duration: req.body.duration,
+      difficulty: req.body.difficulty,
+      image: req.body.image
+    });
+    res.status(201).json({
+      ...exercise,
+      id: exercise.id,
+      name: exercise.name,
+      description: exercise.description,
+      duration: exercise.duration,
+      difficulty: exercise.difficulty,
+      image: exercise.image
+    });
   } catch (err) {
     res.status(500).send(err);
   }
 });
 
-router.delete('/exercises/:id', async (req, res) => {
+router.delete('/exercises/:id', authService.authenticationMiddleware, async (req, res) => {
   try {
     const exerciseDAO: MongoGenericDAO<Exercise> = req.app.locals.exerciseDAO;
     const id = req.params.id;
+    const exercise = await exerciseDAO.findOne({ id });
+    if (!exercise) {
+      res.status(404).send();
+      return;
+    }
+    if (res.locals.user.id !== exercise.userId) {
+      res.status(401).send();
+      return;
+    }
     const success = await exerciseDAO.delete(id);
     if (success) {
       res.status(204).send();
